requirements.txt
asgiref==3.3.1
Django==3.1.6
django-js-asset==1.2.2
django-modeltranslation==0.16.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
